Cate Palmer
Added a welcome message showing the name of the latest user, a list of users who have registered, and a button to sort the list alphabetically (but not the sort function to do this yet). I also used some temporary data in the form of a tempUsers array that I forgot to remove (and remove references to) before submitting this pull request. My Trello board is here: https://trello.com/b/kMRBTMyo/wj4j-cate-palmer